Dedupe getSvgPathFromStroke into src/lib/svg-path.ts
DrawingPad and StrokePreview carried identical copies of the freehand
outline → SVG path conversion; both now import the shared, unit-tested
function.
