+++ src/components/dependent/DrawingPad.tsx
@@ -11,6 +11,7 @@ import { useGetStrokeFn } from "@/hooks/use-get-stroke-fn";
 import { PracticeButton } from "@/components/ui/practice-button";
 import { Undo2, Trash2, Search } from "@/components/icons";
 import { SquareX } from "lucide-react";
+import { getSvgPathFromStroke } from "@/lib/svg-path";
 
 export type Stroke = [number, number][];
 
@@ -19,18 +20,6 @@ export type DrawingSubmitPayload = {
   width: number;
   height: number;
 };
-
-function getSvgPathFromStroke(stroke: number[][]): string {
-  if (!stroke.length) return "";
-  const d: (string | number)[] = ["M", ...stroke[0], "Q"];
-  for (let i = 0; i < stroke.length; i++) {
-    const [x0, y0] = stroke[i];
-    const [x1, y1] = stroke[(i + 1) % stroke.length];
-    d.push(x0, y0, (x0 + x1) / 2, (y0 + y1) / 2);
-  }
-  d.push("Z");
-  return d.join(" ");
-}
 
 export const DrawingPad = ({
   svgSize,
